DE-531 add in ability to calculate yearly aggregate/average data for QA/QC report, refactoring work done for the Trend Report
