+++ eppt-trend-reporting/src/main/java/gov/ca/water/trendreporting/EpptReportingMonths.java
@@ -1,5 +1,5 @@
 /*
- * Enhanced Post Processing Tool (EPPT) Copyright (c) 2019.
+ * Enhanced Post Processing Tool (EPPT) Copyright (c) 2020.
  *
  * EPPT is copyrighted by the State of California, Department of Water Resources. It is licensed
  * under the GNU General Public License, version 2. This means it can be
@@ -16,20 +16,22 @@ import java.nio.file.Paths;
 import java.time.LocalDate;
 import java.time.Month;
 import java.time.YearMonth;
-import java.time.format.DateTimeFormatter;
 import java.time.format.TextStyle;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
+import java.util.NavigableMap;
 import java.util.Objects;
+import java.util.TreeMap;
 import java.util.stream.LongStream;
 
 import gov.ca.water.calgui.EpptInitializationException;
 import gov.ca.water.calgui.bo.CalLiteGUIException;
-import gov.ca.water.calgui.bo.GUILinksAllModelsBO;
 import gov.ca.water.calgui.busservice.impl.GuiLinksSeedDataSvcImpl;
+import gov.ca.water.calgui.busservice.impl.MonthPeriod;
 import gov.ca.water.calgui.constant.Constant;
 import gov.ca.water.calgui.techservice.IFileSystemSvc;
 import gov.ca.water.calgui.techservice.impl.FileSystemSvcImpl;
@@ -47,7 +49,7 @@ import static java.util.stream.Collectors.toList;
  */
 public final class EpptReportingMonths
 {
-	private static final List<MonthPeriod> MONTH_PERIODS = new ArrayList<>();
+	private static final Map<Integer, MonthPeriod> MONTH_PERIODS = new TreeMap<>();
 	private static final Logger LOG = Logger.getLogger(EpptReportingMonths.class.getName());
 	private static final String MONTH_PERIODS_FILENAME = CONFIG_DIR + "trendreporting/TrendReportingPeriod" + CSV_EXT;
 	private static EpptReportingMonths seedDataSvc;
@@ -119,9 +121,8 @@ public final class EpptReportingMonths
 				{
 					end = Month.valueOf(list[2].toUpperCase());
 				}
-				MONTH_PERIODS.add(new MonthPeriod(index, start, end));
+				MONTH_PERIODS.put(index, new MonthPeriod(start, end));
 			}
-			MONTH_PERIODS.sort(Comparator.comparingInt(MonthPeriod::getIndex));
 		}
 		catch(ArrayIndexOutOfBoundsException ex)
 		{
@@ -149,7 +150,7 @@ public final class EpptReportingMonths
 
 	public static List<MonthPeriod> getAllMonthPeriods()
 	{
-		return new ArrayList<>(MONTH_PERIODS);
+		return new ArrayList<>(MONTH_PERIODS.values());
 	}
 
 	public static List<Month> getMonths(MonthPeriod monthPeriod)
@@ -168,110 +169,5 @@ public final class EpptReportingMonths
 						 .limit(between + 1)
 						 .mapToObj(monthPeriod.getStart()::plus)
 						 .collect(toList());
-	}
-
-	public static final class MonthPeriod
-	{
-		private final int _index;
-		private final Month _start;
-		private final Month _end;
-
-		private MonthPeriod(int index, Month start, Month end)
-		{
-			_index = index;
-			_start = start;
-			_end = end;
-		}
-
-		public Month getStart()
-		{
-			return _start;
-		}
-
-		public Month getEnd()
-		{
-			return _end;
-		}
-
-		@Override
-		public String toString()
-		{
-			if(Objects.equals(_start, _end))
-			{
-				return formatMonth(_start);
-			}
-			else
-			{
-				return formatMonth(_start) + " - " + formatMonth(_end);
-			}
-		}
-
-		private String formatMonth(Month month)
-		{
-			if(month != null)
-			{
-				return month.getDisplayName(TextStyle.FULL, Locale.US);
-			}
-			else
-			{
-				return "";
-			}
-		}
-
-		List<YearMonth> getYearMonths(int year)
-		{
-			List<YearMonth> retval = new ArrayList<>();
-			if(_start == null || _end == null)
-			{
-				return retval;
-			}
-			if(_start != _end)
-			{
-				int i = 0;
-				long between = ChronoUnit.MONTHS.between(YearMonth.of(2001, _start), YearMonth.of(2010, _end)) % 12;
-				Month month = _start.plus(i);
-				YearMonth yearMonth;
-				if(isPreviousYear(month))
-				{
-					yearMonth = YearMonth.of(year - 1, month);
-				}
-				else
-				{
-					yearMonth = YearMonth.of(year, month);
-				}
-				while(retval.size() < between + 1)
-				{
-					retval.add(yearMonth.plusMonths(i));
-					i++;
-				}
-			}
-			else
-			{
-				if(isPreviousYear(_start))
-				{
-					retval.add(YearMonth.of(year - 1, _start));
-				}
-				else
-				{
-					retval.add(YearMonth.of(year, _start));
-				}
-			}
-			return retval;
-		}
-
-		private boolean isPreviousYear(Month month)
-		{
-			long totalMonths = 1 + ChronoUnit.MONTHS.between(YearMonth.of(2001, _start), YearMonth.of(2010, _end)) % 12;
-			if(totalMonths == 12 && _start == Month.JANUARY)
-			{
-				return false;
-			}
-			return month == Month.OCTOBER || month == Month.NOVEMBER || month == Month.DECEMBER;
-		}
-
-		public int getIndex()
-		{
-			return _index;
-		}
 	}
 }
