+++ eppt-trend-reporting/src/test/java/gov/ca/water/trendreporting/TrendReportScaffold.java
@@ -16,7 +16,6 @@ import java.awt.BorderLayout;
 import java.awt.Color;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.Collections;
 import javax.swing.*;
